[Refactor]:Order Jpa부분 수정

+++ src/main/java/com/back/_1cafe/order/OrderRepository.java
@@ -2,5 +2,5 @@ package com.back._1cafe.order;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Orders,Long> {
+public interface OrderRepository extends JpaRepository<Orders,Integer> {
 }
